% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_create_report.R
\name{fh_create_report}
\alias{fh_create_report}
\title{Create Report}
\usage{
fh_create_report(
  output_file = "report.html",
  output_dir = file.path(getwd(), "output")
)
}
\arguments{
\item{output_file}{The desired output filename (defaults to "report.html")}

\item{output_dir}{The desired output directory (defaults to "/output")}
}
\value{
TRUE when complete
}
\description{
Create Report
}
